// src/types/section.ts
/** Sections on an author page. */
export type AuthorSection =
  | "about"
  | "works"
  | "search"
  | "summary"
  | "names"
  | "citations"
  | "foreign"
  | "lemmas";

/** Sections on a text page. */
export type TextSection =
  | "title"
  | "content"
  | "search"
  | "summary"
  | "names"
  | "citations"
  | "foreign"
  | "lemmas"
  | "about";